chore(core): remove all send v1 screens
