+++ mobile-app/app/screens/AppNavigator/screens/Portfolio/screens/SendScreen.tsx
@@ -1,59 +1,94 @@
-import { InputHelperText } from '@components/InputHelperText'
-import { WalletTextInput } from '@components/WalletTextInput'
-import { DeFiAddress } from '@defichain/jellyfish-address'
-import { NetworkName } from '@defichain/jellyfish-network'
-import { StackScreenProps } from '@react-navigation/stack'
-import { DFITokenSelector, DFIUtxoSelector, tokensSelector, WalletToken } from '@store/wallet'
+import { useEffect, useState, useMemo, useRef, useCallback } from 'react'
+import { Platform, TextInput, View } from 'react-native'
 import BigNumber from 'bignumber.js'
-import { useCallback, useEffect, useRef, useState } from 'react'
 import { Control, Controller, useForm } from 'react-hook-form'
-import { Platform, View } from 'react-native'
 import { useSelector } from 'react-redux'
-import { AmountButtonTypes, SetAmountButton } from '@components/SetAmountButton'
-import {
-  ThemedIcon,
-  ThemedScrollView,
-  ThemedSectionTitle,
-  ThemedText,
-  ThemedTouchableOpacity,
-  ThemedView
-} from '@components/themed'
+import { useToast } from 'react-native-toast-notifications'
+import { StackScreenProps } from '@react-navigation/stack'
+import { getColor, tailwind } from '@tailwind'
+import { debounce } from 'lodash'
+import { fromAddress } from '@defichain/jellyfish-address'
+import { NetworkName } from '@defichain/jellyfish-network'
+import { translate } from '@translations'
 import { useNetworkContext } from '@shared-contexts/NetworkContext'
 import { useWhaleApiClient } from '@shared-contexts/WhaleContext'
+import { useLogger } from '@shared-contexts/NativeLoggingProvider'
+import { useThemeContext } from '@shared-contexts/ThemeProvider'
 import { RootState } from '@store'
+import { AddressType, DFITokenSelector, DFIUtxoSelector, tokensSelector, WalletToken } from '@store/wallet'
+import { LocalAddress } from '@store/userPreferences'
 import { hasTxQueued as hasBroadcastQueued } from '@store/ocean'
 import { hasTxQueued } from '@store/transaction_queue'
-import { tailwind } from '@tailwind'
-import { translate } from '@translations'
-import { PortfolioParamList } from '../PortfolioNavigator'
-import { InfoRow, InfoType } from '@components/InfoRow'
-import { useLogger } from '@shared-contexts/NativeLoggingProvider'
-import { ConversionInfoText } from '@components/ConversionInfoText'
-import { NumberRow } from '@components/NumberRow'
-import { ReservedDFIInfoText } from '@components/ReservedDFIInfoText'
+import { useDisplayUtxoWarning } from '@hooks/wallet/DisplayUtxoWarning'
 import { queueConvertTransaction, useConversion } from '@hooks/wallet/Conversion'
-import { SymbolIcon } from '@components/SymbolIcon'
-import { BottomSheetModal } from '@gorhom/bottom-sheet'
-import { BottomSheetNavScreen, BottomSheetWebWithNav, BottomSheetWithNav } from '@components/BottomSheetWithNav'
-import { BottomSheetToken, BottomSheetTokenList, TokenType } from '@components/BottomSheetTokenList'
-import { InfoText } from '@components/InfoText'
-import { SubmitButtonGroup } from '@components/SubmitButtonGroup'
-import { LocalAddress } from '@store/userPreferences'
-import { debounce } from 'lodash'
-import { useWalletAddress } from '@hooks/useWalletAddress'
 import { useAppDispatch } from '@hooks/useAppDispatch'
+import { useWalletAddress } from '@hooks/useWalletAddress'
+import {
+  ThemedIcon,
+  ThemedTextInputV2,
+  ThemedTextV2,
+  ThemedTouchableOpacity,
+  ThemedTouchableOpacityV2,
+  ThemedViewV2
+} from '@components/themed'
+import { WalletTextInputV2 } from '@components/WalletTextInputV2'
+import { SubmitButtonGroupV2 } from '@components/SubmitButtonGroupV2'
+import { AmountButtonTypes, TransactionCard, TransactionCardStatus } from '@components/TransactionCard'
+import { useTokenPrice } from '../hooks/TokenPrice'
+import { ActiveUSDValueV2 } from '../../Loans/VaultDetail/components/ActiveUSDValueV2'
+import { PortfolioParamList } from '../PortfolioNavigator'
+import { RandomAvatar } from '../components/RandomAvatar'
+import { TokenIcon } from '../components/TokenIcon'
+import { KeyboardAwareScrollView } from 'react-native-keyboard-aware-scroll-view'
+import { useSafeAreaInsets } from 'react-native-safe-area-context'
 
 type Props = StackScreenProps<PortfolioParamList, 'SendScreen'>
+
+export interface BottomSheetToken {
+  tokenId: string
+  available: BigNumber
+  token: {
+    name: string
+    displaySymbol: string
+    symbol: string
+    isLPS?: boolean
+  }
+  factor?: string
+  reserve?: string
+}
 
 export function SendScreen ({
   route,
   navigation
 }: Props): JSX.Element {
+  const dispatch = useAppDispatch()
   const logger = useLogger()
   const { networkName } = useNetworkContext()
   const client = useWhaleApiClient()
+  const { isLight } = useThemeContext()
+  const { getTokenPrice } = useTokenPrice()
+  const { fetchWalletAddresses } = useWalletAddress()
+  const { getDisplayUtxoWarningStatus } = useDisplayUtxoWarning()
+  const toast = useToast()
+  const TOAST_DURATION = 2000
+  const BOTTOM_NAV_HEIGHT = 64
+  const bottomInset = useSafeAreaInsets().bottom
+
   const tokens = useSelector((state: RootState) => tokensSelector(state.wallet))
+  const addressBook = useSelector((state: RootState) => state.userPreferences.addressBook)
+  const walletAddress = useSelector((state: RootState) => state.userPreferences.addresses)
+  const hasPendingJob = useSelector((state: RootState) => hasTxQueued(state.transactionQueue))
+  const hasPendingBroadcastJob = useSelector((state: RootState) => hasBroadcastQueued(state.ocean))
+  const DFIUtxo = useSelector((state: RootState) => DFIUtxoSelector(state.wallet))
+  const DFIToken = useSelector((state: RootState) => DFITokenSelector(state.wallet))
+
   const [token, setToken] = useState(route.params?.token)
+  const [matchedAddress, setMatchedAddress] = useState<LocalAddress>()
+  const [fee, setFee] = useState<BigNumber>(new BigNumber(0.0001))
+  const [jellyfishWalletAddress, setJellyfishWalletAddresses] = useState<string[]>([])
+  const [transactionCardStatus, setTransactionCardStatus] = useState(TransactionCardStatus.Default)
+
+  // form
   const {
     control,
     setValue,
@@ -63,65 +98,108 @@ export function SendScreen ({
     watch
   } = useForm({ mode: 'onChange' })
   const { address } = watch()
-  const addressBook = useSelector((state: RootState) => state.userPreferences.addressBook)
-  const walletAddress = useSelector((state: RootState) => state.userPreferences.addresses)
-  const [matchedAddress, setMatchedAddress] = useState<LocalAddress>()
-  const dispatch = useAppDispatch()
-  const [fee, setFee] = useState<BigNumber>(new BigNumber(0.0001))
-  const hasPendingJob = useSelector((state: RootState) => hasTxQueued(state.transactionQueue))
-  const hasPendingBroadcastJob = useSelector((state: RootState) => hasBroadcastQueued(state.ocean))
-  const DFIUtxo = useSelector((state: RootState) => DFIUtxoSelector(state.wallet))
-  const DFIToken = useSelector((state: RootState) => DFITokenSelector(state.wallet))
+  const amountToSend = getValues('amount')
+  const [addressType, setAddressType] = useState<AddressType>()
+
   const {
     isConversionRequired,
     conversionAmount
   } = useConversion({
     inputToken: {
       type: token?.id === '0_unified' ? 'utxo' : 'others',
-      amount: new BigNumber(getValues('amount'))
+      amount: new BigNumber(amountToSend)
     },
-    deps: [getValues('amount'), JSON.stringify(token)]
+    deps: [amountToSend, JSON.stringify(token)]
   })
-  const [hasBalance, setHasBalance] = useState(false)
-  const { fetchWalletAddresses } = useWalletAddress()
-  const [jellyfishWalletAddress, setJellyfishWalletAddresses] = useState<string[]>([])
-
-  // Bottom sheet token
-  const [isModalDisplayed, setIsModalDisplayed] = useState(false)
-  const [bottomSheetScreen, setBottomSheetScreen] = useState<BottomSheetNavScreen[]>([])
-  const containerRef = useRef(null)
-  const bottomSheetRef = useRef<BottomSheetModal>(null)
-  const expandModal = useCallback(() => {
-    if (Platform.OS === 'web') {
-      setIsModalDisplayed(true)
-    } else {
-      bottomSheetRef.current?.present()
-    }
-  }, [])
-  const dismissModal = useCallback(() => {
-    if (Platform.OS === 'web') {
-      setIsModalDisplayed(false)
-    } else {
-      bottomSheetRef.current?.close()
-    }
-  }, [])
 
   const debounceMatchAddress = debounce(() => {
     if (address !== undefined && addressBook !== undefined && addressBook[address] !== undefined) {
       setMatchedAddress(addressBook[address])
+      setAddressType(AddressType.Whitelisted)
     } else if (address !== undefined && walletAddress !== undefined && walletAddress[address] !== undefined) {
       setMatchedAddress(walletAddress[address])
+      setAddressType(AddressType.WalletAddress)
     } else if (address !== undefined && jellyfishWalletAddress.includes(address)) {
       // wallet address that does not have a label
       setMatchedAddress({
         address,
-        label: 'Saved address',
+        label: '',
         isMine: true
       })
+      setAddressType(AddressType.WalletAddress)
     } else {
       setMatchedAddress(undefined)
+      setAddressType(fromAddress(address, networkName) !== undefined ? AddressType.OthersButValid : undefined)
     }
   }, 200)
+
+  const reservedDFI = 0.1
+  const isReservedUtxoUsed = getDisplayUtxoWarningStatus(new BigNumber(amountToSend), token?.displaySymbol ?? '')
+
+  const amountInputRef = useRef<TextInput>()
+  const amountInUSDValue = useMemo(() => {
+    if (token === undefined || isNaN(amountToSend) || (amountToSend === '')) {
+      return new BigNumber(0)
+    }
+
+    return getTokenPrice(token.symbol, amountToSend, token.isLPS)
+  }, [amountToSend, token])
+
+  const {
+    infoText,
+    infoTextThemedProps
+  } = useMemo(() => {
+    let infoText
+    let themedProps
+    let status = TransactionCardStatus.Default
+
+    if (new BigNumber(amountToSend).isGreaterThan(token?.amount ?? 0)) {
+      infoText = 'Insufficient balance'
+      themedProps = {
+        dark: tailwind('text-red-v2'),
+        light: tailwind('text-red-v2')
+      }
+      status = TransactionCardStatus.Error
+    } else if (token?.isLPS === true && new BigNumber(amountToSend).isGreaterThan(0)) {
+      infoText = 'Make sure to send your LP Tokens to only DeFiChain-compatible wallets. Failing to do so may lead to irreversible loss of funds'
+      themedProps = {
+        dark: tailwind('text-orange-v2'),
+        light: tailwind('text-orange-v2')
+      }
+    } else if (isReservedUtxoUsed) {
+      infoText = 'A small amount of UTXO is reserved for fees'
+      themedProps = {
+        dark: tailwind('text-orange-v2'),
+        light: tailwind('text-orange-v2')
+      }
+    } else {
+      infoText = 'There is a minimal fee for the transaction'
+      themedProps = {
+        light: tailwind('text-mono-light-v2-500'),
+        dark: tailwind('text-mono-dark-v2-500')
+      }
+    }
+
+    setTransactionCardStatus(status)
+    return {
+      infoText: translate('screens/SendScreen', infoText),
+      infoTextThemedProps: {
+        ...themedProps,
+        style: tailwind('text-xs mt-2 ml-5 font-normal-v2')
+      }
+    }
+  }, [token, isReservedUtxoUsed, amountToSend])
+
+  useEffect(() => {
+    setToken(route.params.token)
+  }, [route.params.token])
+
+  useEffect(() => {
+    /* timeout added to auto display keyboard on Android */
+    Platform.OS === 'android'
+      ? setTimeout(() => amountInputRef?.current?.focus(), 0)
+      : amountInputRef?.current?.focus()
+  }, [])
 
   useEffect(() => {
     void fetchWalletAddresses().then((walletAddresses) => setJellyfishWalletAddresses(walletAddresses))
@@ -136,339 +214,286 @@ export function SendScreen ({
   useEffect(() => {
     const t = tokens.find((t) => t.id === token?.id)
     if (t !== undefined) {
-      setToken({ ...t })
+      setToken({
+        ...t,
+        amount: t.displaySymbol === 'DFI' ? new BigNumber(t.amount).minus(reservedDFI).toFixed(8) : t.amount
+      })
     }
-
-    const totalBalance = tokens.reduce((total, token) => {
-      return total.plus(new BigNumber(token.amount))
-    }, new BigNumber(0))
-
-    setHasBalance(totalBalance.isGreaterThan(0))
   }, [JSON.stringify(tokens)])
 
   useEffect(() => {
     debounceMatchAddress()
   }, [address, addressBook])
 
-  const setTokenListBottomSheet = useCallback(() => {
-    setBottomSheetScreen([
-      {
-        stackScreenName: 'TokenList',
-        component: BottomSheetTokenList({
-          tokens: getBottomSheetToken(tokens),
-          tokenType: TokenType.BottomSheetToken,
-          headerLabel: translate('screens/SendScreen', 'Choose token to send'),
-          onCloseButtonPress: () => dismissModal(),
-          onTokenPress: async (item): Promise<void> => {
-            const _token = tokens.find(t => t.id === item.tokenId)
-            if (_token !== undefined) {
-              setToken(_token)
-              setValue('amount', '')
-              await trigger('amount')
-            }
-            dismissModal()
-          }
-        }),
-        option: {
-          header: () => null
-        }
-      }])
-  }, [])
+  function showToast (type: AmountButtonTypes): void {
+    if (token?.displaySymbol === undefined) {
+      return
+    }
+
+    toast.hideAll()
+    const isMax = type === AmountButtonTypes.Max
+    const toastMessage = isMax ? 'Max available {{unit}} entered' : '{{percent}} of available {{unit}} entered'
+    const toastOption = {
+      unit: token.displaySymbol,
+      percent: type
+    }
+    toast.show(translate('screens/SendScreen', toastMessage, toastOption), {
+      type: 'wallet_toast',
+      placement: 'top',
+      duration: TOAST_DURATION
+    })
+  }
+
+  const onAddressSelect = useCallback(async (savedAddress: string) => {
+    setValue('address', savedAddress, { shouldDirty: true })
+    navigation.goBack()
+    await trigger('address')
+  }, [navigation])
 
   async function onSubmit (): Promise<void> {
-    if (hasPendingJob || hasPendingBroadcastJob || token === undefined) {
+    if (hasPendingJob || hasPendingBroadcastJob || token === undefined || !formState.isValid) {
       return
     }
 
     const values = getValues()
-    if (formState.isValid && isConversionRequired) {
+    const params: PortfolioParamList['SendConfirmationScreen'] = {
+      destination: values.address,
+      token,
+      amount: new BigNumber(values.amount),
+      amountInUsd: amountInUSDValue,
+      fee,
+      toAddressLabel: matchedAddress?.label,
+      addressType
+    }
+
+    if (isConversionRequired) {
       queueConvertTransaction({
         mode: 'accountToUtxos',
         amount: conversionAmount
       }, dispatch, () => {
+        params.conversion = {
+          DFIUtxo,
+          DFIToken,
+          isConversionRequired: true,
+          conversionAmount
+        }
         navigation.navigate({
           name: 'SendConfirmationScreen',
-          params: {
-            destination: values.address,
-            token,
-            amount: new BigNumber(values.amount),
-            amountInUsd: new BigNumber(0), // not displayed in send screen v1
-            fee,
-            conversion: {
-              DFIUtxo,
-              DFIToken,
-              isConversionRequired,
-              conversionAmount
-            }
-          },
+          params,
           merge: true
         })
-      }, logger)
-    } else if (formState.isValid) {
-      const values = getValues()
+      }, logger, () => {
+        params.conversion = {
+          DFIUtxo,
+          DFIToken,
+          isConversionRequired: true,
+          conversionAmount,
+          isConverted: true
+        }
+        navigation.navigate({
+          name: 'SendConfirmationScreen',
+          params,
+          merge: true
+        })
+      })
+    } else {
       navigation.navigate({
         name: 'SendConfirmationScreen',
-        params: {
-          destination: values.address,
-          token,
-          amount: new BigNumber(values.amount),
-          amountInUsd: new BigNumber(0), // not displayed in send screen v1
-          fee
-        },
+        params,
         merge: true
       })
     }
   }
 
+  const onAmountChange = async (amount: string): Promise<void> => {
+    setValue('amount', amount, { shouldDirty: true })
+    await trigger('amount')
+  }
+
   return (
-    <View style={tailwind('h-full')} ref={containerRef}>
-      <ThemedScrollView contentContainerStyle={tailwind('pt-6 pb-8')} testID='send_screen'>
-        <TokenInput
-          token={token}
-          onPress={() => {
-            setTokenListBottomSheet()
-            expandModal()
-          }}
-          isDisabled={!hasBalance}
-        />
+    <View style={tailwind('h-full')}>
+      <KeyboardAwareScrollView
+        contentContainerStyle={tailwind('pt-6 pb-8')} testID='send_screen'
+        style={tailwind(`${isLight ? 'bg-mono-light-v2-100' : 'bg-mono-dark-v2-100'}`)}
+        extraScrollHeight={-BOTTOM_NAV_HEIGHT - bottomInset}
+      >
+        {token === undefined &&
+          <ThemedTextV2 style={tailwind('px-5')}>
+            {translate('screens/SendScreen', 'Select a token you want to send to get started')}
+          </ThemedTextV2>}
 
-        {token === undefined
-          ? (
-            <ThemedText style={tailwind('px-4')}>
-              {translate('screens/SendScreen', 'Select a token you want to send to get started')}
-            </ThemedText>
-          )
-          : (
-            <>
-              <View style={tailwind('px-4')}>
-                <AddressRow
-                  control={control}
-                  networkName={networkName}
-                  onContactButtonPress={() => navigation.navigate({
-                    name: 'AddressBookScreen',
-                    params: {
-                      selectedAddress: getValues('address'),
-                      onAddressSelect: (savedAddres: string) => {
-                        setValue('address', savedAddres, { shouldDirty: true })
-                        navigation.goBack()
-                      }
-                    },
-                    merge: true
-                  })}
-                  onQrButtonPress={() => navigation.navigate({
-                    name: 'BarCodeScanner',
-                    params: {
-                      onQrScanned: async (value) => {
-                        setValue('address', value, { shouldDirty: true })
-                        await trigger('address')
-                      }
-                    },
-                    merge: true
-                  })}
-                  onClearButtonPress={async () => {
-                    setValue('address', '')
-                    await trigger('address')
-                  }}
-                  onAddressChange={async (address) => {
-                    setValue('address', address, { shouldDirty: true })
-                    await trigger('address')
-                  }}
-                  inputFooter={
-                    <>
-                      {matchedAddress !== undefined && (
-                        <ThemedView
-                          style={tailwind('mx-2 mb-2 p-1 rounded-2xl flex flex-row self-start', { 'items-end': Platform.OS === 'ios' })}
-                          light={tailwind('bg-gray-50')}
-                          dark={tailwind('bg-gray-900')}
-                        >
-                          <ThemedIcon
-                            name='account-check'
-                            iconType='MaterialCommunityIcons'
-                            size={18}
-                            light={tailwind('text-gray-400')}
-                            dark={tailwind('text-gray-500')}
-                          />
-                          <ThemedText
-                            style={tailwind('text-xs ml-1 pt-px')}
-                            light={tailwind('text-gray-500')}
-                            dark={tailwind('text-gray-400')}
-                            testID='address_input_footer'
-                          >
-                            {matchedAddress.label}
-                          </ThemedText>
-                        </ThemedView>
-                      )}
-                    </>
+        {token !== undefined && (
+          <View style={tailwind('px-5')}>
+            <View style={tailwind('my-12 items-center')}>
+              <Controller
+                control={control}
+                defaultValue=''
+                name='amount'
+                render={({
+                  field: {
+                    onChange,
+                    value
                   }
-                />
-                <AmountRow
-                  control={control}
-                  onAmountChange={async (amount) => {
-                    setValue('amount', amount, { shouldDirty: true })
-                    await trigger('amount')
-                  }}
-                  onClearButtonPress={async () => {
-                    setValue('amount', '')
-                    await trigger('amount')
-                  }}
-                  token={token}
-                />
+                }) => (
+                  <ThemedTextInputV2
+                    style={tailwind('text-3xl text-center font-semibold-v2 w-full')}
+                    light={tailwind('text-mono-light-v2-900')}
+                    dark={tailwind('text-mono-dark-v2-900')}
+                    keyboardType='numeric'
+                    value={value}
+                    onChange={onChange}
+                    onChangeText={onAmountChange}
+                    placeholder='0.00'
+                    placeholderTextColor={getColor(isLight ? 'mono-light-v2-900' : 'mono-dark-v2-900')}
+                    testID='amount_input'
+                    ref={amountInputRef}
+                  />
+                )}
+                rules={{
+                  required: true,
+                  pattern: /^\d*\.?\d*$/,
+                  max: BigNumber.max(token.amount, 0).toFixed(8),
+                  validate: {
+                    greaterThanZero: (value: string) => new BigNumber(value !== undefined && value !== '' ? value : 0).isGreaterThan(0)
+                  }
+                }}
+              />
+              <ActiveUSDValueV2
+                price={amountInUSDValue} testId='amount_input_in_usd'
+                containerStyle={tailwind('w-full break-words')}
+                style={tailwind('text-center justify-center w-full')}
+              />
+            </View>
 
-                <ReservedDFIInfoText />
-                {isConversionRequired &&
-                  <View style={tailwind('mt-2')}>
-                    <ConversionInfoText />
-                  </View>}
-              </View>
-              {
-                fee !== undefined && (
-                  <View style={tailwind()}>
-                    <ThemedSectionTitle
-                      text={translate('screens/SendScreen', 'TRANSACTION DETAILS')}
-                    />
-                    {isConversionRequired &&
-                      <NumberRow
-                        lhs={translate('screens/SendScreen', 'UTXO to be converted')}
-                        rhs={{
-                          value: conversionAmount.toFixed(8),
-                          testID: 'text_amount_to_convert',
-                          suffixType: 'text',
-                          suffix: token.displaySymbol
-                        }}
-                      />}
+            <AmountCard
+              onPress={async () => {
+                navigation.navigate({
+                  name: 'TokenSelectionScreen',
+                  params: {},
+                  merge: true
+                })
+                setValue('amount', '', { shouldDirty: true })
+                await trigger('amount')
+              }}
+              onAmountChange={async (amount: string, type: AmountButtonTypes) => {
+                showToast(type)
+                setValue('amount', amount, { shouldDirty: true })
+                await trigger('amount')
+              }}
+              token={token}
+              transactionCardStatus={transactionCardStatus}
+            />
+            <ThemedTextV2 {...infoTextThemedProps} testID='info_text'>{infoText}</ThemedTextV2>
 
-                    <InfoRow
-                      type={InfoType.EstimatedFee}
-                      value={fee.toString()}
-                      testID='transaction_fee'
-                      suffix='DFI'
+            <AddressRow
+              control={control}
+              networkName={networkName}
+              onContactButtonPress={() => navigation.navigate({
+                name: 'AddressBookScreen',
+                params: {
+                  selectedAddress: getValues('address'),
+                  onAddressSelect
+                },
+                merge: true
+              })}
+              onQrButtonPress={() => navigation.navigate({
+                name: 'BarCodeScanner',
+                params: {
+                  onQrScanned: async (value) => {
+                    setValue('address', value, { shouldDirty: true })
+                    await trigger('address')
+                  }
+                },
+                merge: true
+              })}
+              onClearButtonPress={async () => {
+                setValue('address', '')
+                await trigger('address')
+              }}
+              onAddressChange={async (address) => {
+                setValue('address', address, { shouldDirty: true })
+                await trigger('address')
+              }}
+            />
+
+            <View style={tailwind('ml-5 my-2 items-center flex flex-row')}>
+              {addressType === AddressType.OthersButValid
+                ? (
+                  <>
+                    <ThemedIcon
+                      light={tailwind('text-success-500')}
+                      dark={tailwind('text-darksuccess-500')}
+                      iconType='MaterialIcons'
+                      name='check-circle'
+                      size={16}
                     />
-                  </View>
+                    <ThemedTextV2
+                      style={tailwind('text-xs mx-1 font-normal-v2')}
+                      light={tailwind('text-mono-light-v2-500')}
+                      dark={tailwind('text-mono-dark-v2-500')}
+                    >
+                      {translate('screens/SendScreen', 'Verified')}
+                    </ThemedTextV2>
+                  </>
                 )
-              }
-              <ThemedText
-                testID='transaction_details_info_text'
-                light={tailwind('text-gray-600')}
-                dark={tailwind('text-gray-300')}
-                style={tailwind('mt-2 mx-4 text-sm')}
-              >
-                {isConversionRequired
-                  ? translate('screens/SendScreen', 'Authorize transaction in the next screen to convert')
-                  : translate('screens/SendScreen', 'Review full transaction details in the next screen')}
-              </ThemedText>
-            </>
-          )}
+                : (addressType !== undefined && (
+                  <ThemedViewV2
+                    style={tailwind('flex flex-row items-center overflow-hidden rounded-lg py-0.5', {
+                        'px-1': addressType === AddressType.WalletAddress,
+                        'px-2': addressType === AddressType.Whitelisted
+                      })}
+                    light={tailwind('bg-mono-light-v2-200')}
+                    dark={tailwind('bg-mono-dark-v2-200')}
+                  >
+                    {addressType === AddressType.WalletAddress && (
+                      <View style={tailwind('rounded-l-2xl mr-1')}>
+                        <RandomAvatar name={matchedAddress?.address} size={12} />
+                      </View>
+                      )}
 
-        <View style={tailwind('mt-6')}>
-          <SubmitButtonGroup
+                    <ThemedTextV2
+                      ellipsizeMode='middle'
+                      numberOfLines={1}
+                      style={[tailwind('text-xs font-normal-v2'), {
+                          minWidth: 10,
+                          maxWidth: 108
+                        }]}
+                      light={tailwind('text-mono-light-v2-500')}
+                      dark={tailwind('text-mono-dark-v2-500')}
+                      testID='address_input_footer'
+                    >
+                      {matchedAddress?.label !== '' ? matchedAddress?.label : matchedAddress.address}
+                    </ThemedTextV2>
+                  </ThemedViewV2>
+                  )
+                )}
+            </View>
+
+          </View>
+        )}
+
+        <View style={tailwind('mt-24 mx-12 items-center')}>
+          {(formState.isValid && token !== undefined) &&
+            <ThemedTextV2
+              testID='transaction_details_info_text'
+              light={tailwind('text-mono-light-v2-500')}
+              dark={tailwind('text-mono-dark-v2-500')}
+              style={tailwind('mt-2 text-xs text-center font-normal-v2')}
+            >
+              {isConversionRequired
+                ? translate('screens/SendScreen', 'By continuing, the required amount of DFI will be converted')
+                : translate('screens/SendScreen', 'Review full details in the next screen')}
+            </ThemedTextV2>}
+          <SubmitButtonGroupV2
             isDisabled={!formState.isValid || hasPendingJob || hasPendingBroadcastJob || token === undefined}
-            label={translate('screens/SendScreen', 'CONTINUE')}
-            processingLabel={translate('screens/SendScreen', 'CONTINUE')}
+            label={translate('screens/SendScreen', 'Continue')}
             onSubmit={onSubmit}
             title='send_continue'
-            isProcessing={hasPendingJob || hasPendingBroadcastJob}
             displayCancelBtn={false}
+            buttonStyle='mt-5'
           />
         </View>
-
-        {Platform.OS === 'web' && (
-          <BottomSheetWebWithNav
-            modalRef={containerRef}
-            screenList={bottomSheetScreen}
-            isModalDisplayed={isModalDisplayed}
-            modalStyle={{
-              position: 'absolute',
-              height: '350px',
-              width: '375px',
-              zIndex: 50,
-              bottom: '0'
-            }}
-          />
-        )}
-
-        {Platform.OS !== 'web' && (
-          <BottomSheetWithNav
-            modalRef={bottomSheetRef}
-            screenList={bottomSheetScreen}
-          />
-        )}
-      </ThemedScrollView>
-    </View>
-  )
-}
-
-function TokenInput (props: { token?: WalletToken, onPress: () => void, isDisabled: boolean }): JSX.Element {
-  const hasNoBalanceForSelectedToken = props.token?.amount === undefined ? true : new BigNumber(props.token?.amount).lte(0)
-  return (
-    <View style={tailwind('px-4')}>
-      <ThemedText
-        style={tailwind('text-xl font-semibold')}
-      >
-        {translate('screens/SendScreen', 'Send to other wallet')}
-      </ThemedText>
-      <ThemedTouchableOpacity
-        onPress={props.onPress}
-        dark={tailwind({
-          'bg-gray-600 text-gray-500 border-0': props.isDisabled,
-          'border-gray-600 bg-gray-800': !props.isDisabled
-        })}
-        light={tailwind({
-          'bg-gray-200 border-0': props.isDisabled,
-          'border-gray-300 bg-white': !props.isDisabled
-        })}
-        style={tailwind('border rounded w-full flex flex-row justify-between h-12 items-center px-2', {
-          'mb-10': props.token?.isLPS === false,
-          'mb-2': props.token?.isLPS === true,
-          'mb-6': props.token === undefined
-        })}
-        testID='select_token_input'
-        disabled={props.isDisabled}
-      >
-        {props.token === undefined || props.isDisabled || hasNoBalanceForSelectedToken
-          ? (
-            <ThemedText
-              light={tailwind('text-gray-500')}
-              dark={tailwind('text-gray-400')}
-              style={tailwind('text-sm')}
-              testID='select_token_placeholder'
-            >
-              {translate('screens/SendScreen', 'Select token')}
-            </ThemedText>
-          )
-          : (
-            <View style={tailwind('flex flex-row')}>
-              <SymbolIcon symbol={props.token.displaySymbol} styleProps={tailwind('w-6 h-6')} />
-              <ThemedText
-                style={tailwind('ml-2 font-medium')}
-                testID='selected_token'
-              >
-                {props.token.displaySymbol}
-              </ThemedText>
-            </View>
-          )}
-        <ThemedIcon
-          iconType='MaterialIcons'
-          name='unfold-more'
-          size={24}
-          dark={tailwind({
-            'text-darkprimary-500': !props.isDisabled,
-            'text-gray-400': props.isDisabled
-          })}
-          light={tailwind({
-            'text-primary-500': !props.isDisabled,
-            'text-gray-500': props.isDisabled
-          })}
-          style={tailwind('-mr-1.5 flex-shrink-0')}
-        />
-      </ThemedTouchableOpacity>
-      {props.token?.isLPS === true &&
-        (
-          <InfoText
-            testID='lp_info_text'
-            text={translate('components/ConversionInfoText', 'Send Liquidity Pool tokens only to DeFiChain compatible wallets. Otherwise, sending to other exchanges may result in irreversible loss of funds.')}
-            style={tailwind('mb-10')}
-          />
-        )}
+      </KeyboardAwareScrollView>
     </View>
   )
 }
@@ -484,181 +509,171 @@ function AddressRow ({
 }: { control: Control, networkName: NetworkName, onContactButtonPress: () => void, onQrButtonPress: () => void, onClearButtonPress: () => void, onAddressChange: (address: string) => void, inputFooter?: React.ReactElement }): JSX.Element {
   const defaultValue = ''
   return (
-    <>
-      <Controller
-        control={control}
-        defaultValue={defaultValue}
-        name='address'
-        render={({
-          field: {
-            value,
-            onChange
-          }
-        }) => (
-          <View style={tailwind('flex-row w-full')}>
-            <WalletTextInput
+    <Controller
+      control={control}
+      defaultValue={defaultValue}
+      name='address'
+      render={({
+        field: {
+          value,
+          onChange
+        },
+        fieldState: { error }
+      }) => {
+        const hasValidAddress = error?.type == null
+        return (
+          <View style={tailwind('flex w-full')}>
+            <WalletTextInputV2
               autoCapitalize='none'
               multiline
-              onChange={onChange}
-              onChangeText={onAddressChange}
-              placeholder={translate('screens/SendScreen', 'Paste wallet address here')}
-              style={tailwind('w-3/5 flex-grow pb-1')}
+              onBlur={async () => {
+                await onAddressChange(value?.trim())
+              }}
+              onChangeText={onChange}
+              placeholder={translate('screens/SendScreen', 'Paste address')}
+              style={tailwind('w-3/5 flex-grow pb-1 font-normal-v2')}
               testID='address_input'
               value={value}
               displayClearButton={value !== defaultValue}
               onClearButtonPress={onClearButtonPress}
-              title={translate('screens/SendScreen', 'Where do you want to send?')}
+              title={translate('screens/SendScreen', 'SEND TO')}
               titleTestID='title_to_address'
               inputType='default'
               inputFooter={inputFooter}
+              valid={hasValidAddress}
             >
-              <ThemedTouchableOpacity
-                dark={tailwind('bg-gray-800 border-gray-400')}
-                light={tailwind('bg-white border-gray-300')}
-                onPress={onContactButtonPress}
-                style={tailwind('w-9 p-1.5 mr-1 border rounded')}
-                testID='address_book_button'
+              {value !== '' && <View style={tailwind('mr-2')} />}
+              {value === '' &&
+                <>
+                  <ThemedTouchableOpacity
+                    dark={tailwind('bg-black')}
+                    light={tailwind('bg-white')}
+                    onPress={onContactButtonPress}
+                    style={tailwind('w-9 p-1.5 rounded')}
+                    testID='address_book_button'
+                  >
+                    <ThemedIcon
+                      iconType='Feather'
+                      dark={tailwind('text-mono-dark-v2-700')}
+                      light={tailwind('text-mono-light-v2-700')}
+                      name='users'
+                      size={24}
+                    />
+                  </ThemedTouchableOpacity>
+                  <ThemedTouchableOpacity
+                    dark={tailwind('bg-black')}
+                    light={tailwind('bg-white')}
+                    onPress={onQrButtonPress}
+                    style={tailwind('w-9 p-1.5 rounded')}
+                    testID='qr_code_button'
+                  >
+                    <ThemedIcon
+                      dark={tailwind('text-mono-dark-v2-700')}
+                      light={tailwind('text-mono-light-v2-700')}
+                      iconType='MaterialIcons'
+                      name='qr-code'
+                      size={24}
+                    />
+                  </ThemedTouchableOpacity>
+                </>}
+            </WalletTextInputV2>
+            {/* TODO: Replace with inline comment if possible @pierregee */}
+            {/* TODO: Update with required error message also */}
+            {!hasValidAddress &&
+              <ThemedTextV2
+                style={tailwind('text-xs mt-2 ml-5 font-normal-v2')}
+                dark={tailwind('text-red-v2')}
+                light={tailwind('text-red-v2')}
+                testID='address_error_text'
               >
-                <ThemedIcon
-                  dark={tailwind('text-darkprimary-500')}
-                  iconType='MaterialCommunityIcons'
-                  light={tailwind('text-primary-500')}
-                  name='account-multiple'
-                  size={24}
-                />
-              </ThemedTouchableOpacity>
-              <ThemedTouchableOpacity
-                dark={tailwind('bg-gray-800 border-gray-400')}
-                light={tailwind('bg-white border-gray-300')}
-                onPress={onQrButtonPress}
-                style={tailwind('w-9 p-1.5 border rounded')}
-                testID='qr_code_button'
-              >
-                <ThemedIcon
-                  dark={tailwind('text-darkprimary-500')}
-                  iconType='MaterialIcons'
-                  light={tailwind('text-primary-500')}
-                  name='qr-code-scanner'
-                  size={24}
-                />
-              </ThemedTouchableOpacity>
-            </WalletTextInput>
+                {translate('screens/SendScreen', 'Invalid address. Make sure the address is correct to avoid irrecoverable losses')}
+              </ThemedTextV2>}
           </View>
-        )}
-        rules={{
-          required: true,
-          validate: {
-            isValidAddress: (address) => DeFiAddress.from(networkName, address).valid
-          }
-        }}
-      />
-    </>
+        )
+      }}
+      rules={{
+        required: true,
+        validate: {
+          isValidAddress: (address) => (fromAddress(address, networkName) !== undefined)
+        }
+      }}
+    />
   )
 }
 
 interface AmountForm {
-  control: Control
+  transactionCardStatus: TransactionCardStatus
   token: WalletToken
-  onAmountChange: (amount: string) => void
-  onClearButtonPress: () => void
+  onPress: () => void
+  onAmountChange: (amount: string, type: AmountButtonTypes) => Promise<void>
 }
 
-function AmountRow ({
+function AmountCard ({
+  transactionCardStatus,
   token,
-  control,
-  onAmountChange,
-  onClearButtonPress
+  onPress,
+  onAmountChange
 }: AmountForm): JSX.Element {
-  const reservedDFI = 0.1
-  let maxAmount = token.symbol === 'DFI' ? new BigNumber(token.amount).minus(reservedDFI).toFixed(8) : token.amount
-  maxAmount = BigNumber.max(maxAmount, 0).toFixed(8)
-  const defaultValue = ''
+  const maxAmount = BigNumber.max(token.amount, 0)
   return (
     <>
-      <Controller
-        control={control}
-        defaultValue={defaultValue}
-        name='amount'
-        render={({
-          field: {
-            onChange,
-            value
-          }
-        }) => (
-          <ThemedView
-            dark={tailwind('bg-transparent')}
-            light={tailwind('bg-transparent')}
-            style={tailwind('flex-row w-full mt-6')}
-          >
-            <WalletTextInput
-              autoCapitalize='none'
-              onChange={onChange}
-              onChangeText={onAmountChange}
-              placeholder={translate('screens/SendScreen', 'Enter an amount')}
-              style={tailwind('flex-grow w-2/5')}
-              testID='amount_input'
-              value={value}
-              displayClearButton={value !== defaultValue}
-              onClearButtonPress={onClearButtonPress}
-              title={translate('screens/SendScreen', 'How much do you want to send?')}
-              titleTestID='title_send'
-              inputType='numeric'
-            >
-              <ThemedView
-                dark={tailwind('bg-gray-800')}
-                light={tailwind('bg-white')}
-                style={tailwind('flex-row items-center')}
-              >
-                <SetAmountButton
-                  amount={new BigNumber(maxAmount)}
-                  onPress={onAmountChange}
-                  type={AmountButtonTypes.half}
-                />
-
-                <SetAmountButton
-                  amount={new BigNumber(maxAmount)}
-                  onPress={onAmountChange}
-                  type={AmountButtonTypes.max}
-                />
-              </ThemedView>
-            </WalletTextInput>
-
-          </ThemedView>
-        )}
-        rules={{
-          required: true,
-          pattern: /^\d*\.?\d*$/,
-          max: maxAmount,
-          validate: {
-            greaterThanZero: (value: string) => new BigNumber(value !== undefined && value !== '' ? value : 0).isGreaterThan(0)
-          }
-        }}
-      />
-
-      <InputHelperText
-        testID='max_value'
-        label={`${translate('screens/SendScreen', 'Available')}: `}
-        content={maxAmount}
-        suffix={` ${token.displaySymbol}`}
-      />
+      <ThemedTextV2
+        style={tailwind('pl-5 pb-2 text-xs font-normal-v2')}
+        light={tailwind('text-mono-light-v2-500')}
+        dark={tailwind('text-mono-dark-v2-500')}
+      >{translate('screens/SendScreen', 'I WANT TO SEND')}
+      </ThemedTextV2>
+      <TransactionCard
+        status={transactionCardStatus}
+        maxValue={maxAmount}
+        onChange={onAmountChange}
+        containerStyle={tailwind('rounded-t-lg-v2')}
+      >
+        <ThemedTouchableOpacityV2
+          style={tailwind('flex flex-row items-center justify-between pt-4.5 mb-4 mx-5')}
+          onPress={onPress}
+          testID='select_token_input'
+        >
+          <View style={tailwind('flex flex-row items-center')}>
+            <TokenIcon
+              testID={`${token.displaySymbol}_icon`}
+              token={{
+                isLPS: token.isLPS,
+                displaySymbol: token.displaySymbol
+              }}
+              size={32}
+              iconBStyle={tailwind('-ml-3')}
+            />
+            <View style={tailwind('flex ml-2')}>
+              <ThemedTextV2>
+                <ThemedTextV2
+                  style={tailwind('font-semibold-v2 text-sm')}
+                  testID='max_value'
+                >{maxAmount.toFixed(8)}
+                </ThemedTextV2>
+                <ThemedTextV2
+                  style={tailwind('font-semibold-v2 text-sm')}
+                  testID='max_value_display_symbol'
+                >{` ${token.displaySymbol}`}
+                </ThemedTextV2>
+              </ThemedTextV2>
+              <ThemedTextV2
+                light={tailwind('text-mono-light-v2-500')}
+                dark={tailwind('text-mono-dark-v2-500')}
+                style={tailwind('text-xs font-normal-v2')}
+              >{translate('screens/SendScreen', 'Available')}
+              </ThemedTextV2>
+            </View>
+          </View>
+          <ThemedIcon
+            dark={tailwind('text-mono-dark-v2-700')}
+            light={tailwind('text-mono-light-v2-700')}
+            iconType='Feather'
+            name='chevron-right'
+            size={28}
+          />
+        </ThemedTouchableOpacityV2>
+      </TransactionCard>
     </>
   )
-}
-
-function getBottomSheetToken (tokens: WalletToken[]): BottomSheetToken[] {
-  return tokens.filter(t => {
-    return new BigNumber(t.amount).isGreaterThan(0) && t.id !== '0' && t.id !== '0_utxo'
-  }).map(t => {
-    const token: BottomSheetToken = {
-      tokenId: t.id,
-      available: new BigNumber(t.amount),
-      token: {
-        name: t.name,
-        displaySymbol: t.displaySymbol,
-        symbol: t.symbol,
-        isLPS: t.isLPS
-      }
-    }
-    return token
-  })
 }
